+++ pages/news.md
@@ -1,10 +1,13 @@
 ---
+name: Recent Updates
 header_image: /images/ukelele-2.jpg
 header_text: Recent Updates
 eleventyNavigation:
   key: News
   order: 3
-layout: news-archive.html
 permalink: /news/
+blocks:
+  - type: markdown
+    content: |
+      # Check out our latest news posts below
 ---
-# Check out our latest news posts below